requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
flake8==7.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
mccabe==0.7.0
mypy_extensions==1.1.0
openai==1.76.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
python-dotenv==1.1.0
requests==2.32.3
slack_bolt==1.23.0
slack_sdk==3.35.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0